chore: move commands pkg to internal root
