+++ internal/eventhandlers/interaction_create.go
@@ -3,8 +3,8 @@ package eventhandlers
 import (
 	"github.com/bwmarrin/discordgo"
 	"github.com/elliotwms/pinbot/internal/commandhandlers"
+	"github.com/elliotwms/pinbot/internal/commands"
 	"github.com/elliotwms/pinbot/internal/config"
-	"github.com/elliotwms/pinbot/internal/pinbot/commands"
 	"github.com/sirupsen/logrus"
 )
